create/requirements.txt
python_telegram_bot==13.11
requests==2.27.1
